adjust typings, make sure cached users never have sequelize wrapping around them
