+++ package.json
@@ -13,7 +13,8 @@
     "build": "nuxt-ts build && backpack build",
     "start": "cross-env NODE_ENV=production node build/main.js",
     "start-elasticbeanstalk": "NODE_ENV=production node build/main.js",
-    "generate": "nuxt-ts generate"
+    "generate": "nuxt-ts generate",
+    "postinstall": "patch-package"
   },
   "dependencies": {
     "axios": "^0.21.1",
@@ -39,6 +40,7 @@
     "morgan": "~1.10.0",
     "nuxt": "^2.14.6",
     "nuxt-property-decorator": "^2.8.8",
+    "patch-package": "^6.4.7",
     "pg": "^8.3.3",
     "sass-loader": "^10.0.2",
     "sequelize": "^6.3.5",
