+++ types/express.d.ts
@@ -1,3 +1,5 @@
+import { switchShareUserToken } from './flintgg';
+
 declare namespace Express {
   export interface Request {
     user: switchShareUserToken;
