+++ types/flintgg.ts
@@ -1,0 +1,14 @@
+type Bigint = string; // later, when fully supported: type = bigint
+
+export type flintId = Bigint; // bigint
+
+export type flintError = {
+  title: string;
+  detail: string;
+  status: number;
+};
+
+export type switchShareUserToken = {
+  id: string;
+  name: string;
+};
